refactor: colonizethis_turn wave-5 slice A support densify (#4168) (#4169)
* refactor: colonizethis_turn wave-5 slice A support densify (#4168)

Add shared scenario builders to turn_resolver_test_harness, migrate six
integration scenario suites to kRegion* constants and harness helpers,
introduce repo.turn_test_support_region_literals gate, and lower the
support LOC ratchet ceiling to 5610.

Refs #4168


* refactor: complete wave-5 slice A support densify (#4168)

Finish densifying combat_resolution_continued, spy/fog, and dialogue
integration scenarios via shared harness builders; lower support LOC
ratchet ceiling to 5675 (5475 physical LOC, 200-line headroom).

Refs #4168


* fix(turn): unblock slice A CI and densify combat logging tests (#4168)

Exclude colonizethis_turn test/support from the cross-domain 400-line
test cap (governed by repo.turn_test_support_loc) and extract combat
logging phase fixtures into a sibling cases module per wave-5 slice B.


* refactor(turn): complete wave-5 slice B root test densify (#4168)

Extract shared fixtures into *_cases.dart siblings for research funding,
order projections, and economy stockpile preview suites so each root test
file is at or below 280 physical lines per wave-5 AC 5.

Refs #4168


* refactor: colonizethis_turn wave-5 slice C lib module splits (#4168)

Split spy_resolver, naval_resolution_move, turn_news_digest, and
world_market_phase_orders into focused sibling modules while preserving
public exports and characterization test behavior.

Refs #4168


* fix: retarget allProvinces sanctions after turn_news_digest split

Slice C moved province scans into turn_news_digest_discovery.dart and
turn_news_digest_terrestrial.dart; update sanctions YAML to match.


---------
