+++ SPEC/program/repo-lint.md
@@ -66,7 +66,7 @@
 | `tool/check_logic_test_file_size.dart` | Enforce `packages/colonizethis_logic/test/**` Dart files at **400 physical lines or fewer** (physical line count); rule `repo.logic_test_file_size` runs **full-tree** in `ct_repo_lint` and narrows to changed files when an incremental baseline is supplied (`CT_REPO_LINT_BASE_SHA` in CI). Originally PR-incremental only against the #2216 debt; transitioned to full-tree enforcement once the residual `test/**` violations were cleared (GitHub #2288). The dedicated `quality.yml` step for changed logic tests remains as a fast early signal on PRs. |
 | `tool/check_logic_test_tree_loc.dart` | Enforce total physical LOC under `packages/colonizethis_logic/test/**` at **≤5,800** (post-#4090 orphan purge ratchet); rule `repo.logic_test_tree_loc`. |
 | `tool/check_logic_test_orphan_basenames.dart` | Forbid reintroduction of the thirteen purged domain collision basenames under `packages/colonizethis_logic/test/**` (Refs #4090 AC3/AC5); rule `repo.logic_test_orphan_basenames`. |
-| `tool/check_domain_package_test_file_size.dart` | Enforce **400 physical lines or fewer** across every `.dart` file under `packages/colonizethis_<domain>/test/**` for the eight split logic-domain packages (`world`, `combat`, `economy`, `diplomacy`, `setup`, `orders`, `turn`, `ai_contracts`); rule `repo.domain_package_test_file_size` (Refs #3290). Repository test `test/check_domain_package_test_file_size_test.dart` asserts the full-tree scan stays green on `dev`. |
+| `tool/check_domain_package_test_file_size.dart` | Enforce **400 physical lines or fewer** across every `.dart` file under `packages/colonizethis_<domain>/test/**` for the eight split logic-domain packages (`world`, `combat`, `economy`, `diplomacy`, `setup`, `orders`, `turn`, `ai_contracts`), excluding `packages/colonizethis_turn/test/support/**` (governed by `repo.turn_test_support_loc` / `repo.turn_test_file_size`); rule `repo.domain_package_test_file_size` (Refs #3290). Repository test `test/check_domain_package_test_file_size_test.dart` asserts the full-tree scan stays green on `dev`. |
 | `tool/check_dart_file_non_comment_line_size.dart` | Enforce repository-wide Dart files at **1000 non-comment lines or fewer** (fail when strictly greater), excluding generated suffixes and generated l10n path prefixes — see `SPEC/program/dart-file-non-comment-line-size.md`; rule `repo.dart_file_non_comment_line_size` |
 | `tool/check_long_string_switches.dart` | Warn/fail on large string-literal switch statements/expressions (warn ≥20, fail ≥50) via manifest rule `repo.dart_long_string_switches`; uses shared repo-wide collector `collectRepoLintRepoWideDartFiles` — see `SPEC/program/show-tech-tool.md` |
 | `tool/check_land_province_bucket_keys.dart` | For guarded explore/fog/news paths, disallow local-only land-province tile-bucket lookups (`tileKeysByRegionAndProvince[region]?[localId]`); require canonical full-id buckets only; rule `repo.land_province_bucket_keys` |
@@ -119,7 +119,8 @@
 | `tool/check_turn_world_market_test_support.dart` | World-market tests under `packages/colonizethis_turn/test/turn/` that invoke `worldMarketTurnPhaseHandler` directly MUST import `test/support/world_market_test_support.dart` and call a shared runner (`runWorldMarketPhase` / `runWorldMarketPhasePipeline` / `From` variants, or wrappers `runWorldMarketFrrCreditPhase` / `runTreasuryClampPhase` / `runWorldMarketTradePhase`). Phase-handler suites (`*_phase_*_test.dart`) must use a runner even when they never spell the handler symbol. Allowlisted pure helper unit suites: `world_market_phase_orders_helpers_test.dart`, `world_market_phase_completed_trade_pairs_test.dart`, `world_market_phase_deals_subsidy_test.dart` (Refs #3876 / #4039); rule `repo.turn_world_market_test_support` |
 | `tool/check_turn_test_resolve_complete.dart` | Forbid nested `requireTurnResolutionComplete(resolveTurnForGame(...))` and direct `resolveTurnForGame(` under `packages/colonizethis_turn/test/**` outside the harness and a short pending-result allowlist; success paths MUST use `resolveTurnComplete` from `test/support/turn_resolver_test_harness.dart` (Refs #4039); rule `repo.turn_test_resolve_complete` |
 | `tool/check_turn_test_fixture_location.dart` | Forbid `*_fixtures.dart` / `*_test_support.dart` under `packages/colonizethis_turn/test/**` outside `test/support/` so fixture modules stay co-located with harnesses (Refs #4039); rule `repo.turn_test_fixture_location` |
-| `tool/check_turn_test_support_loc.dart` | Ratchet ceiling for all Dart under `packages/colonizethis_turn/test/support/` (physical LOC; wave-4 post-densify ceiling **5800**); fail on net regrowth beyond the ceiling (Refs #4039, #4113); rule `repo.turn_test_support_loc` |
+| `tool/check_turn_test_support_loc.dart` | Ratchet ceiling for all Dart under `packages/colonizethis_turn/test/support/` (physical LOC; wave-5 slice A post-densify ceiling **5675**); fail on net regrowth beyond the ceiling (Refs #4039, #4113, #4168); rule `repo.turn_test_support_loc` |
+| `tool/check_turn_test_support_region_literals.dart` | Forbid bare `'oldWorld'` / `'newWorld'` string literals under `packages/colonizethis_turn/test/support/**` (use `kRegionOldWorld` / `kRegionNewWorld`); rule `repo.turn_test_support_region_literals` (Refs #4168) |
 | `tool/check_turn_lib_file_size.dart` | Ratchet ceiling for `packages/colonizethis_turn/lib/**` physical lines per file (wave-4 ≤400; shrink-only grandfather allowlist during transition; generated suffixes excluded); rule `repo.turn_lib_file_size` (Refs #4113) |
 | `tool/check_turn_test_file_size.dart` | Ratchet ceiling for `packages/colonizethis_turn/test/**` physical lines per file excluding `test/support/` (wave-4 ≤400; support tree governed by `repo.turn_test_support_loc`); rule `repo.turn_test_file_size` (Refs #4113) |
 | `tool/check_world_lib_file_size.dart` | Ratchet ceiling for `packages/colonizethis_world/lib/**` physical lines per file (wave-5 ≤400; empty shrink-only grandfather after slice D; generated suffixes excluded); rule `repo.colonizethis_world_lib_file_size` (Refs #4125) |
@@ -385,7 +386,9 @@ Large `app/test/**` widget/screen suites that are split into numbered fragments 
 
 **Diplomacy test support file size (Refs #4130):** `repo.colonizethis_diplomacy_test_support_file_size` fails when any non-generated Dart file under `packages/colonizethis_diplomacy_test_support/lib/**` exceeds **400** physical lines outside the shrink-only grandfather allowlist (empty after wave-5 slice E densify).
 
-**Turn test support LOC (Refs #4039, #4113):** `repo.turn_test_support_loc` fails when physical LOC of all Dart under `packages/colonizethis_turn/test/support/` exceeds the documented post-densify ceiling (**5800**). Ratchet downward as densify slices shrink the tree.
+**Turn test support LOC (Refs #4039, #4113, #4168):** `repo.turn_test_support_loc` fails when physical LOC of all Dart under `packages/colonizethis_turn/test/support/` exceeds the documented post-densify ceiling (**5675** after wave-5 slice A completion). Ratchet downward as densify slices shrink the tree.
+
+**Turn test support region literals (Refs #4168):** `repo.turn_test_support_region_literals` fails when any non-comment line under `packages/colonizethis_turn/test/support/**` contains a bare `'oldWorld'` or `'newWorld'` string literal; use `kRegionOldWorld` / `kRegionNewWorld` from `package:colonizethis_logic/colonizethis_logic.dart`.
 
 **Feature-screen top-bar helper (Refs #3952):** Diplomacy / Production / Technology / Trade screens build Map-back + 18×18 icon + title chrome via `app/lib/widgets/game_feature_screen_top_bar.dart` (`GameFeatureScreenTopBar.build`); screen-specific `topBarKey` / title / icon asset constants remain on each screen.
