+++ packages/colonizethis_turn/test/research_phase_funding_test.dart
@@ -1,80 +1,49 @@
 import 'package:colonizethis_test/test.dart';
 import 'package:colonizethis_data/colonizethis_data.dart';
-import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
+import 'research_phase_funding_cases.dart';
 import 'support/turn_game_fixtures.dart';
-import 'support/turn_resolver_test_harness.dart';
 
 void main() {
   group('Research phase funding', () {
     test('accumulates research progress and unlocks tech when cost reached', () {
       final tech = techById(kTechIdCropRotation)!;
-      // Maximum funding costs 1000 gold/turn (per SPEC/game/tech-tree.md)
       final game = researchPhaseTestBaseGame(
         treasury: 2000,
         techUnlocked: const {},
       );
-
-      final orders = Orders(
-        researchOrdersByPlayerId: {
-          'p1': const [
-            ResearchOrder(
-              slotIndex: 0,
-              techId: kTechIdCropRotation,
-              funding: ResearchFundingLevel.maximum,
-            ),
-          ],
-        },
+      final orders = p1ResearchOrders(
+        techId: kTechIdCropRotation,
+        funding: ResearchFundingLevel.maximum,
       );
 
-      final topology = const MapTopology();
-      final next = resolveTurnComplete(game: game, topology: topology, orders: orders);
-      final player = next.players.single;
+      final player = resolveResearchPlayer(game: game, orders: orders);
 
-      // One turn of maximum funding should make progress > 0 and reduce treasury.
       expect(player.treasury, lessThan(game.players.single.treasury));
       final progress = player.researchProgressByTechId ?? const {};
       final unlocked = player.techUnlocked ?? const {};
 
       if (progress.isNotEmpty) {
-        // Research not yet complete; progress should be less than cost and tech not unlocked.
         expect(progress[kTechIdCropRotation], isNotNull);
         expect(progress[kTechIdCropRotation], lessThan(tech.cost));
         expect(unlocked[kTechIdCropRotation], isNot(true));
       } else {
-        // If cost is small, research may complete in a single turn.
         expect(unlocked[kTechIdCropRotation], isTrue);
       }
     });
 
     test('applies prerequisite rule: cannot research tech without prereqs', () {
-      // No spend when prereq not met; use enough treasury in case logic ever applied
       final game = researchPhaseTestBaseGame(
         treasury: 2000,
-        techUnlocked: const {}, // saw_mill not researched
+        techUnlocked: const {},
       );
-
-      final orders = Orders(
-        researchOrdersByPlayerId: const {
-          'p1': [
-            ResearchOrder(
-              slotIndex: 0,
-              techId: kTechIdWindSawMill,
-              funding: ResearchFundingLevel.maximum,
-            ),
-          ],
-        },
+      final orders = p1ResearchOrders(
+        techId: kTechIdWindSawMill,
+        funding: ResearchFundingLevel.maximum,
       );
+      final player = resolveResearchPlayer(game: game, orders: orders);
 
-      final next = resolveTurnComplete(
-          game: game,
-          topology: const MapTopology(),
-          orders: orders,
-        );
-      final player = next.players.single;
-
-      // Treasury unchanged and no progress recorded because prerequisite not met.
       expect(player.treasury, game.players.single.treasury);
       expect(player.researchProgressByTechId ?? const {}, isEmpty);
       expect(player.techUnlocked?[kTechIdWindSawMill], isNot(true));
@@ -87,58 +56,31 @@ void main() {
           treasury: 100,
           techUnlocked: const {},
         );
-        final orders = Orders(
-          researchOrdersByPlayerId: {
-            'p1': const [
-              ResearchOrder(
-                slotIndex: 0,
-                techId: kTechIdCropRotation,
-                funding: ResearchFundingLevel.none,
-              ),
-            ],
-          },
+        final orders = p1ResearchOrders(
+          techId: kTechIdCropRotation,
+          funding: ResearchFundingLevel.none,
         );
-        final next = resolveTurnComplete(
-            game: game,
-            topology: const MapTopology(),
-            orders: orders,
-          );
-        expect(next.players.single.treasury, 100);
-        expect(
-          next.players.single.researchProgressByTechId ?? const {},
-          isEmpty,
-        );
+        final player = resolveResearchPlayer(game: game, orders: orders);
+
+        expect(player.treasury, 100);
+        expect(player.researchProgressByTechId ?? const {}, isEmpty);
       },
     );
 
     test('research with low funding deducts treasury and adds progress', () {
-      // Use wind_saw_mill (tier-2 cost 2400) with prereq so 100 RP does not
-      // complete in one turn. SPEC/game/tech-tree.md § Research Model.
       final game = researchPhaseTestBaseGame(
         treasury: 100,
         techUnlocked: const {kTechIdSawMill: true},
       );
-      final orders = Orders(
-        researchOrdersByPlayerId: {
-          'p1': const [
-            ResearchOrder(
-              slotIndex: 0,
-              techId: kTechIdWindSawMill,
-              funding: ResearchFundingLevel.low,
-            ),
-          ],
-        },
+      final orders = p1ResearchOrders(
+        techId: kTechIdWindSawMill,
+        funding: ResearchFundingLevel.low,
       );
-      final next = resolveTurnComplete(
-          game: game,
-          topology: const MapTopology(),
-          orders: orders,
-        );
-      // Low funding: 50 gold cost, 100 RP per turn (per SPEC/game/tech-tree.md)
-      expect(next.players.single.treasury, 50);
+      final player = resolveResearchPlayer(game: game, orders: orders);
+
+      expect(player.treasury, 50);
       expect(
-        (next.players.single.researchProgressByTechId ??
-            const {})[kTechIdWindSawMill],
+        (player.researchProgressByTechId ?? const {})[kTechIdWindSawMill],
         100,
       );
     });
@@ -148,213 +90,80 @@ void main() {
         treasury: 2000,
         techUnlocked: const {},
       );
-      final orders = Orders(
-        researchOrdersByPlayerId: {
-          'p1': const [
-            ResearchOrder(
-              slotIndex: 0,
-              techId: kTechIdCropRotation,
-              funding: ResearchFundingLevel.maximum,
-            ),
-          ],
-        },
+      final orders = p1ResearchOrders(
+        techId: kTechIdCropRotation,
+        funding: ResearchFundingLevel.maximum,
       );
-      final next = resolveTurnComplete(
-          game: game,
-          topology: const MapTopology(),
-          orders: orders,
-        );
-      // Maximum funding: 1000 gold cost, 2500 RP per turn (2.5x efficiency).
-      // crop_rotation tier-1 cost is 1800, so 2500 RP >= 1800 unlocks the tech
-      // and progress is cleared. SPEC/game/tech-tree.md § Research Model.
-      expect(next.players.single.treasury, 1000);
-      expect(next.players.single.techUnlocked![kTechIdCropRotation], isTrue);
+      final player = resolveResearchPlayer(game: game, orders: orders);
+
+      expect(player.treasury, 1000);
+      expect(player.techUnlocked![kTechIdCropRotation], isTrue);
     });
 
     test(
       'debt cap rejects allocation that would breach max debt (no money lending)',
       () {
-        // maxDebt = 0 without Money Lending: treasury 30, low funding cost 50
-        // would reach -20, below the floor, so the order is rejected with no
-        // treasury spend and no progress. Refs #3416 (research context).
         final game = researchPhaseTestBaseGame(
           treasury: 30,
           techUnlocked: const {kTechIdSawMill: true},
         );
-        final orders = Orders(
-          researchOrdersByPlayerId: {
-            'p1': const [
-              ResearchOrder(
-                slotIndex: 0,
-                techId: kTechIdWindSawMill,
-                funding: ResearchFundingLevel.low,
-              ),
-            ],
-          },
+        final orders = p1ResearchOrders(
+          techId: kTechIdWindSawMill,
+          funding: ResearchFundingLevel.low,
         );
-        final next = resolveTurnComplete(
-            game: game,
-            topology: const MapTopology(),
-            orders: orders,
-          );
-        expect(next.players.single.treasury, 30);
-        expect(
-          next.players.single.researchProgressByTechId ?? const {},
-          isEmpty,
-        );
-        expect(
-          next.players.single.techUnlocked?[kTechIdWindSawMill],
-          isNot(true),
-        );
+        final player = resolveResearchPlayer(game: game, orders: orders);
+
+        expect(player.treasury, 30);
+        expect(player.researchProgressByTechId ?? const {}, isEmpty);
+        expect(player.techUnlocked?[kTechIdWindSawMill], isNot(true));
       },
     );
 
     test(
       'debt cap allows treasury to go negative within money-lending floor',
       () {
-        // maxDebt = 500 with Money Lending: treasury 30, low funding cost 50
-        // reaches -20 which is within the floor, so the order applies.
         final game = researchPhaseTestBaseGame(
           treasury: 30,
           techUnlocked: const {kTechIdSawMill: true, kTechIdMoneyLending: true},
         );
-        final orders = Orders(
-          researchOrdersByPlayerId: {
-            'p1': const [
-              ResearchOrder(
-                slotIndex: 0,
-                techId: kTechIdWindSawMill,
-                funding: ResearchFundingLevel.low,
-              ),
-            ],
-          },
+        final orders = p1ResearchOrders(
+          techId: kTechIdWindSawMill,
+          funding: ResearchFundingLevel.low,
         );
-        final next = resolveTurnComplete(
-            game: game,
-            topology: const MapTopology(),
-            orders: orders,
-          );
-        expect(next.players.single.treasury, -20);
+        final player = resolveResearchPlayer(game: game, orders: orders);
+
+        expect(player.treasury, -20);
         expect(
-          (next.players.single.researchProgressByTechId ??
-              const {})[kTechIdWindSawMill],
+          (player.researchProgressByTechId ?? const {})[kTechIdWindSawMill],
           100,
         );
       },
     );
 
     test('all funding levels match spec values via game behavior', () {
-      // Use wind_saw_mill (tier-2 cost 2400) with prereq met. Only Maximum
-      // funding (2500 RP) completes it in one turn; Low/Medium/High accrue
-      // partial progress. SPEC/game/tech-tree.md § Research Model.
-      const prereqMet = {kTechIdSawMill: true};
-
-      // Low: 50 gold, 100 RP (no unlock; 100 < 2400)
-      var game = researchPhaseTestBaseGame(
-        treasury: 100,
-        techUnlocked: prereqMet,
-      );
-      var orders = Orders(
-        researchOrdersByPlayerId: {
-          'p1': const [
-            ResearchOrder(
-              slotIndex: 0,
-              techId: kTechIdWindSawMill,
-              funding: ResearchFundingLevel.low,
-            ),
-          ],
-        },
-      );
-      var next = resolveTurnComplete(
-          game: game,
-          topology: const MapTopology(),
-          orders: orders,
+      for (final scenario in windSawMillFundingLevelScenarios) {
+        final game = researchPhaseTestBaseGame(
+          treasury: scenario.treasury,
+          techUnlocked: windSawMillPrereqMet,
         );
-      expect(next.players.single.treasury, 50);
-      expect(
-        (next.players.single.researchProgressByTechId ??
-            const {})[kTechIdWindSawMill],
-        100,
-      );
-
-      // Medium: 150 gold, 300 RP (no unlock; 300 < 2400)
-      game = researchPhaseTestBaseGame(treasury: 200, techUnlocked: prereqMet);
-      orders = Orders(
-        researchOrdersByPlayerId: {
-          'p1': const [
-            ResearchOrder(
-              slotIndex: 0,
-              techId: kTechIdWindSawMill,
-              funding: ResearchFundingLevel.medium,
-            ),
-          ],
-        },
-      );
-      next = resolveTurnComplete(
-          game: game,
-          topology: const MapTopology(),
-          orders: orders,
+        final orders = p1ResearchOrders(
+          techId: kTechIdWindSawMill,
+          funding: scenario.funding,
         );
-      expect(next.players.single.treasury, 50);
-      expect(
-        (next.players.single.researchProgressByTechId ??
-            const {})[kTechIdWindSawMill],
-        300,
-      );
-      expect(
-        next.players.single.techUnlocked?[kTechIdWindSawMill],
-        isNot(true),
-      );
+        final player = resolveResearchPlayer(game: game, orders: orders);
 
-      // High: 400 gold, 800 RP (no unlock; 800 < 2400)
-      game = researchPhaseTestBaseGame(treasury: 500, techUnlocked: prereqMet);
-      orders = Orders(
-        researchOrdersByPlayerId: {
-          'p1': const [
-            ResearchOrder(
-              slotIndex: 0,
-              techId: kTechIdWindSawMill,
-              funding: ResearchFundingLevel.high,
-            ),
-          ],
-        },
-      );
-      next = resolveTurnComplete(
-          game: game,
-          topology: const MapTopology(),
-          orders: orders,
-        );
-      expect(next.players.single.treasury, 100);
-      expect(
-        (next.players.single.researchProgressByTechId ??
-            const {})[kTechIdWindSawMill],
-        800,
-      );
-      expect(
-        next.players.single.techUnlocked?[kTechIdWindSawMill],
-        isNot(true),
-      );
-
-      // Maximum: 1000 gold, 2500 RP (2.5x efficiency, unlocks)
-      game = researchPhaseTestBaseGame(treasury: 1500, techUnlocked: prereqMet);
-      orders = Orders(
-        researchOrdersByPlayerId: {
-          'p1': const [
-            ResearchOrder(
-              slotIndex: 0,
-              techId: kTechIdWindSawMill,
-              funding: ResearchFundingLevel.maximum,
-            ),
-          ],
-        },
-      );
-      next = resolveTurnComplete(
-          game: game,
-          topology: const MapTopology(),
-          orders: orders,
-        );
-      expect(next.players.single.treasury, 500);
-      expect(next.players.single.techUnlocked![kTechIdWindSawMill], isTrue);
+        expect(player.treasury, scenario.expectedTreasury);
+        if (scenario.unlocked) {
+          expect(player.techUnlocked![kTechIdWindSawMill], isTrue);
+        } else {
+          expect(
+            (player.researchProgressByTechId ??
+                const {})[kTechIdWindSawMill],
+            scenario.expectedProgress,
+          );
+          expect(player.techUnlocked?[kTechIdWindSawMill], isNot(true));
+        }
+      }
     });
   });
 }
