+++ packages/colonizethis_turn/test/support/integration/resolve_turn_combat_resolution_continued_scenarios.dart
@@ -10,8 +10,8 @@ void registerCombatResolutionContinuedTests() {
       test(
         'validateOrdersAndResolveTurn filters invalid order and applies only valid move',
         () {
-          final topology = twoAdjacentOldWorldProvinceTopology();
           const ow = turnTestOldWorldRegionId;
+          final topology = twoAdjacentOldWorldProvinceTopology();
           final game = adjacentOwP1P2Game(
             province1OwnerId: 'p1',
             province2OwnerId: 'p1',
@@ -23,10 +23,10 @@ void registerCombatResolutionContinuedTests() {
                 locationProvinceId: '$ow|P1',
               ),
             ],
-            playerVisibilityByTile: const {
+            playerVisibilityByTile: {
               'p1': {
-                'oldWorld|P1|0|0': 'fullyVisible',
-                'oldWorld|P2|0|0': 'fullyVisible',
+                turnTestOwTileKey('P1'): 'fullyVisible',
+                turnTestOwTileKey('P2'): 'fullyVisible',
               },
             },
             players: const [Player(id: 'p1', displayName: 'P1', isHuman: true)],
@@ -34,8 +34,14 @@ void registerCombatResolutionContinuedTests() {
           final orders = Orders(
             moveOrdersByPlayerId: {
               'p1': [
-                MoveOrder(unitId: 'u1', destinationTileKey: '$ow|P2|0|0'),
-                MoveOrder(unitId: 'u999', destinationTileKey: '$ow|P2|0|0'),
+                MoveOrder(
+                  unitId: 'u1',
+                  destinationTileKey: turnTestOwTileKey('P2'),
+                ),
+                MoveOrder(
+                  unitId: 'u999',
+                  destinationTileKey: turnTestOwTileKey('P2'),
+                ),
               ],
             },
           );
@@ -58,160 +64,62 @@ void registerCombatResolutionContinuedTests() {
       );
 
       test('movement phase applies naval mission order', () {
-        final topology = MapTopology(
-          nodes: const [
-            TopologyNode(
-              id: 'sea1',
-              regionId: 'oldWorld',
-              type: TopologyNodeType.seaZone,
-            ),
-          ],
-          edges: const [],
-        );
-        final game = Game(
-          id: 'g1',
-          worldState: WorldState(
-            turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 0),
-            oldWorld: const RegionData(),
-            newWorld: const RegionData(),
-            fleets: [
-              Fleet(
-                id: 'f1',
-                ownerId: 'p1',
-                seaZoneId: 'sea1',
-                regionId: 'oldWorld',
-                shipTypeIds: ['carrack'],
-                mission: FleetMission.none,
-              ),
-            ],
-          ),
-          players: const [Player(id: 'p1', displayName: 'A', isHuman: true)],
-        );
-        final orders = Orders(
-          navalMissionOrdersByPlayerId: {
-            'p1': [
-              NavalMissionOrder(
-                fleetId: 'f1',
-                mission: FleetMission.patrol.name,
-              ),
-            ],
-          },
-        );
         final next = resolveTurnComplete(
-          game: game,
-          topology: topology,
-          orders: orders,
-          extractedByPlayerId: const {},
-          defaultAssignments: const [],
+          game: turnTestFleetsOnlyGame(
+            fleets: [turnTestCarrackFleet(mission: FleetMission.none)],
+          ),
+          topology: turnTestOwSeaZoneTopology(),
+          orders: Orders(
+            navalMissionOrdersByPlayerId: {
+              'p1': [
+                NavalMissionOrder(
+                  fleetId: 'f1',
+                  mission: FleetMission.patrol.name,
+                ),
+              ],
+            },
+          ),
         );
         expect(next.worldState.fleets.single.mission, FleetMission.patrol);
         expect(next.worldState.turnState.turnNumber, 1);
       });
 
       test('movement phase applies naval move order to adjacent sea zone', () {
-        final topology = MapTopology(
-          nodes: const [
-            TopologyNode(
-              id: 'sea1',
-              regionId: 'oldWorld',
-              type: TopologyNodeType.seaZone,
-            ),
-            TopologyNode(
-              id: 'sea2',
-              regionId: 'oldWorld',
-              type: TopologyNodeType.seaZone,
-            ),
-          ],
-          edges: const [TopologyEdge(id1: 'sea1', id2: 'sea2')],
-        );
-        final game = Game(
-          id: 'g1',
-          worldState: WorldState(
-            turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 0),
-            oldWorld: const RegionData(),
-            newWorld: const RegionData(),
-            fleets: [
-              Fleet(
-                id: 'f1',
-                ownerId: 'p1',
-                seaZoneId: 'sea1',
-                regionId: 'oldWorld',
-                shipTypeIds: ['carrack'],
-              ),
-            ],
-          ),
-          players: const [Player(id: 'p1', displayName: 'A', isHuman: true)],
-        );
-        final orders = Orders(
-          navalMoveOrdersByPlayerId: {
-            'p1': [NavalMoveOrder(fleetId: 'f1', destinationSeaZoneId: 'sea2')],
-          },
-        );
         final next = resolveTurnComplete(
-          game: game,
-          topology: topology,
-          orders: orders,
-          extractedByPlayerId: const {},
-          defaultAssignments: const [],
+          game: turnTestFleetsOnlyGame(fleets: [turnTestCarrackFleet()]),
+          topology: turnTestOwTwoLinkedSeaZonesTopology(),
+          orders: Orders(
+            navalMoveOrdersByPlayerId: {
+              'p1': [
+                NavalMoveOrder(fleetId: 'f1', destinationSeaZoneId: 'sea2'),
+              ],
+            },
+          ),
         );
         expect(next.worldState.fleets.single.seaZoneId, 'sea2');
         expect(next.worldState.turnState.turnNumber, 1);
       });
 
       test('dock order moves fleet from sea to port at owned province', () {
-        const ow = 'oldWorld';
-        final topology = MapTopology(
-          nodes: const [
-            TopologyNode(
-              id: 'sea1',
-              regionId: ow,
-              type: TopologyNodeType.seaZone,
-            ),
-            TopologyNode(
-              id: 'P1',
-              regionId: ow,
-              type: TopologyNodeType.province,
-            ),
-          ],
-          edges: const [TopologyEdge(id1: 'sea1', id2: 'P1')],
-        );
-        final game = Game(
-          id: 'g1',
-          worldState: WorldState(
-            turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 0),
-            oldWorld: RegionData(
-              provinces: [Province(id: '$ow|P1', regionId: ow, ownerId: 'p1')],
-            ),
-            newWorld: const RegionData(),
-            fleets: [
-              Fleet(
-                id: 'f1',
-                ownerId: 'p1',
-                seaZoneId: 'sea1',
-                inPortAtProvinceId: null,
-                regionId: ow,
-                shipTypeIds: const ['carrack'],
-              ),
-            ],
-          ),
-          players: const [Player(id: 'p1', displayName: 'A', isHuman: true)],
-        );
-        final orders = Orders(
-          navalMoveOrdersByPlayerId: {
-            'p1': [
-              NavalMoveOrder(
-                fleetId: 'f1',
-                destinationPortProvinceId: '$ow|P1',
-              ),
-            ],
-          },
-        );
+        const ow = kRegionOldWorld;
         final next = resolveTurnComplete(
-          game: game,
-          topology: topology,
-          orders: orders,
-          extractedByPlayerId: const {},
-          defaultAssignments: const [],
+          game: turnTestOwGame(
+            provinces: [
+              Province(id: '$ow|P1', regionId: ow, ownerId: 'p1'),
+            ],
+            fleets: [turnTestCarrackFleet()],
+          ),
+          topology: turnTestOwSeaProvinceTopology(),
+          orders: Orders(
+            navalMoveOrdersByPlayerId: {
+              'p1': [
+                NavalMoveOrder(
+                  fleetId: 'f1',
+                  destinationPortProvinceId: '$ow|P1',
+                ),
+              ],
+            },
+          ),
         );
         final fleet = next.worldState.fleets.single;
         expect(fleet.isInPort, isTrue);
@@ -220,56 +128,27 @@ void registerCombatResolutionContinuedTests() {
       });
 
       test('naval move order undocks fleet from port to adjacent sea zone', () {
-        const ow = 'oldWorld';
-        final topology = MapTopology(
-          nodes: const [
-            TopologyNode(
-              id: 'sea1',
-              regionId: ow,
-              type: TopologyNodeType.seaZone,
-            ),
-            TopologyNode(
-              id: 'P1',
-              regionId: ow,
-              type: TopologyNodeType.province,
-            ),
-          ],
-          edges: const [TopologyEdge(id1: 'sea1', id2: 'P1')],
-        );
-        final game = Game(
-          id: 'g1',
-          worldState: WorldState(
-            turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 0),
-            oldWorld: RegionData(
-              provinces: [Province(id: '$ow|P1', regionId: ow, ownerId: 'p1')],
-            ),
-            newWorld: const RegionData(),
+        const ow = kRegionOldWorld;
+        final next = resolveTurnComplete(
+          game: turnTestOwGame(
+            provinces: [
+              Province(id: '$ow|P1', regionId: ow, ownerId: 'p1'),
+            ],
             fleets: [
-              Fleet(
-                id: 'f1',
-                ownerId: 'p1',
+              turnTestCarrackFleet(
                 seaZoneId: null,
                 inPortAtProvinceId: '$ow|P1',
-                regionId: ow,
-                shipTypeIds: const ['carrack'],
               ),
             ],
           ),
-          players: const [Player(id: 'p1', displayName: 'A', isHuman: true)],
-        );
-        final orders = Orders(
-          navalMoveOrdersByPlayerId: {
-            'p1': [
-              const NavalMoveOrder(fleetId: 'f1', destinationSeaZoneId: 'sea1'),
-            ],
-          },
-        );
-        final next = resolveTurnComplete(
-          game: game,
-          topology: topology,
-          orders: orders,
-          extractedByPlayerId: const {},
-          defaultAssignments: const [],
+          topology: turnTestOwSeaProvinceTopology(),
+          orders: Orders(
+            navalMoveOrdersByPlayerId: {
+              'p1': [
+                const NavalMoveOrder(fleetId: 'f1', destinationSeaZoneId: 'sea1'),
+              ],
+            },
+          ),
         );
         final fleet = next.worldState.fleets.single;
         expect(fleet.isAtSea, isTrue);
