+++ packages/colonizethis_turn/test/support/integration/resolve_turn_combat_resolution_scenarios.dart
@@ -210,64 +210,39 @@ void registerCombatResolutionTests() {
       test(
         'naval interception combat with AI players invokes onDialogue with event battle_won/battle_lost',
         () {
-          final topology = MapTopology(
-            nodes: const [
-              TopologyNode(
-                id: 'sea1',
-                regionId: 'oldWorld',
-                type: TopologyNodeType.seaZone,
+          final dialogueEvents = <DialogueEvent>[];
+          final next = resolveTurnComplete(
+            game: turnTestOwGame(
+              provinces: const [],
+            fleets: [
+              turnTestCarrackFleet(
+                shipTypeIds: const ['carrack', 'carrack'],
               ),
-            ],
-            edges: const [],
-          );
-          final game = Game(
-            id: 'g1',
-            globalGameSeed: 42,
-            worldState: WorldState(
-              turnState: const TurnState(
-                phase: TurnPhase.orders,
-                turnNumber: 0,
-              ),
-              oldWorld: const RegionData(),
-              newWorld: const RegionData(),
-              fleets: [
-                Fleet(
-                  id: 'f1',
-                  ownerId: 'p1',
-                  seaZoneId: 'sea1',
-                  regionId: 'oldWorld',
-                  shipTypeIds: ['carrack', 'carrack'],
-                ),
                 Fleet(
                   id: 'f2',
                   ownerId: 'p2',
                   seaZoneId: 'sea1',
-                  regionId: 'oldWorld',
+                  regionId: kRegionOldWorld,
                   shipTypeIds: ['fluyte'],
                 ),
               ],
-            ),
-            players: const [
-              Player(id: 'p1', displayName: 'AI Fleet A', isHuman: false),
-              Player(id: 'p2', displayName: 'AI Fleet B', isHuman: false),
-            ],
-            diplomacyRelations: [
-              DiplomacyRelation(
-                factionId1: 'p1',
-                factionId2: 'p2',
-                state: RelationState.atWar,
-              ),
-            ],
-          );
-          final dialogueEvents = <DialogueEvent>[];
-          final next = resolveTurnComplete(
-            game: game,
-            topology: topology,
+              players: const [
+                Player(id: 'p1', displayName: 'AI Fleet A', isHuman: false),
+                Player(id: 'p2', displayName: 'AI Fleet B', isHuman: false),
+              ],
+              diplomacyRelations: [
+                DiplomacyRelation(
+                  factionId1: 'p1',
+                  factionId2: 'p2',
+                  state: RelationState.atWar,
+                ),
+              ],
+            ).copyWith(globalGameSeed: 42),
+            topology: turnTestOwSeaZoneTopology(),
             orders: const Orders(),
             eventSink: TurnEventSink(onDialogue: dialogueEvents.add),
           );
           expect(next.worldState.turnState.turnNumber, 1);
-          // Naval battle may or may not eliminate one side; when it does, event dialogue is emitted.
           final eventDialogue = dialogueEvents
               .where(
                 (e) =>
