+++ packages/colonizethis_turn/test/support/integration/resolve_turn_diplomacy_victory_scenarios.dart
@@ -10,60 +10,34 @@ void registerDiplomacyVictoryCoreTests() {
       test(
         'join_home_fleet mission moves ships into home fleet and removes fleet',
         () {
-          const ow = 'oldWorld';
-          final topology = MapTopology(
-            nodes: const [
-              TopologyNode(
-                id: 'sea1',
-                regionId: ow,
-                type: TopologyNodeType.seaZone,
-              ),
-              TopologyNode(
-                id: 'P1',
-                regionId: ow,
-                type: TopologyNodeType.province,
-              ),
-            ],
-            edges: const [TopologyEdge(id1: 'sea1', id2: 'P1')],
-          );
-          final capitalId = '$ow|P1';
-          final homeFleet = Fleet(
+          final topology = turnTestOwSeaProvinceTopology();
+          final capitalId = turnTestOwProvinceId('P1');
+          final homeFleet = turnTestCarrackFleet(
             id: 'fleet_p1',
-            ownerId: 'p1',
             seaZoneId: null,
             inPortAtProvinceId: capitalId,
-            regionId: ow,
-            shipTypeIds: const ['carrack'],
           );
-          final otherFleet = Fleet(
+          final otherFleet = turnTestCarrackFleet(
             id: 'f2',
-            ownerId: 'p1',
             seaZoneId: null,
             inPortAtProvinceId: capitalId,
-            regionId: ow,
             shipTypeIds: const ['fluyte'],
           );
-          final game = Game(
-            id: 'g1',
-            worldState: WorldState(
-              turnState: const TurnState(
-                phase: TurnPhase.orders,
-                turnNumber: 0,
+          final game = turnTestOwGame(
+            provinces: [
+              Province(
+                id: capitalId,
+                regionId: kRegionOldWorld,
+                ownerId: 'p1',
               ),
-              oldWorld: RegionData(
-                provinces: [
-                  Province(id: capitalId, regionId: ow, ownerId: 'p1'),
-                ],
-              ),
-              newWorld: const RegionData(),
-              fleets: [homeFleet, otherFleet],
-            ),
+            ],
+            fleets: [homeFleet, otherFleet],
             players: [
-              const Player(
+              Player(
                 id: 'p1',
                 displayName: 'A',
                 isHuman: true,
-                capitalProvinceId: 'oldWorld|P1',
+                capitalProvinceId: capitalId,
               ),
             ],
           );
@@ -100,45 +74,10 @@ void registerDiplomacyVictoryCoreTests() {
       test(
         'blockade order not applied when not at war with province owner',
         () {
-          const ow = 'oldWorld';
-          final topology = MapTopology(
-            nodes: const [
-              TopologyNode(
-                id: 'sea1',
-                regionId: ow,
-                type: TopologyNodeType.seaZone,
-              ),
-            ],
-            edges: const [],
-          );
-          final game = Game(
-            id: 'g1',
-            worldState: WorldState(
-              turnState: const TurnState(
-                phase: TurnPhase.orders,
-                turnNumber: 0,
-              ),
-              oldWorld: RegionData(
-                provinces: [
-                  Province(id: '$ow|P1', regionId: ow, ownerId: 'p1'),
-                  Province(id: '$ow|P2', regionId: ow, ownerId: 'p2'),
-                ],
-              ),
-              newWorld: const RegionData(),
-              fleets: [
-                Fleet(
-                  id: 'f1',
-                  ownerId: 'p1',
-                  seaZoneId: 'sea1',
-                  regionId: ow,
-                  shipTypeIds: const ['carrack'],
-                ),
-              ],
-            ),
-            players: const [
-              Player(id: 'p1', displayName: 'A', isHuman: true),
-              Player(id: 'p2', displayName: 'B', isHuman: true),
-            ],
+          final topology = turnTestOwSeaZoneTopology();
+          final game = turnTestOwGame(
+            provinces: turnTestOwP1P2Provinces(),
+            fleets: [turnTestCarrackFleet()],
             diplomacyRelations: [
               DiplomacyRelation(
                 factionId1: 'p1',
@@ -153,7 +92,7 @@ void registerDiplomacyVictoryCoreTests() {
                 NavalMissionOrder(
                   fleetId: 'f1',
                   mission: FleetMission.blockade.name,
-                  targetProvinceId: '$ow|P2',
+                  targetProvinceId: turnTestOwProvinceId('P2'),
                 ),
               ],
             },
@@ -171,43 +110,19 @@ void registerDiplomacyVictoryCoreTests() {
       );
 
       test('existing blockade cleared when not at war with target owner', () {
-        const ow = 'oldWorld';
-        final topology = MapTopology(
-          nodes: const [
-            TopologyNode(
-              id: 'sea1',
-              regionId: ow,
-              type: TopologyNodeType.seaZone,
+        final topology = turnTestOwSeaZoneTopology();
+        final game = turnTestOwGame(
+          provinces: turnTestOwP1P2Provinces(),
+          fleets: [
+            Fleet(
+              id: 'f1',
+              ownerId: 'p1',
+              seaZoneId: 'sea1',
+              regionId: kRegionOldWorld,
+              mission: FleetMission.blockade,
+              targetProvinceId: turnTestOwProvinceId('P2'),
+              shipTypeIds: const ['carrack'],
             ),
-          ],
-          edges: const [],
-        );
-        final game = Game(
-          id: 'g1',
-          worldState: WorldState(
-            turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 0),
-            oldWorld: RegionData(
-              provinces: [
-                Province(id: '$ow|P1', regionId: ow, ownerId: 'p1'),
-                Province(id: '$ow|P2', regionId: ow, ownerId: 'p2'),
-              ],
-            ),
-            newWorld: const RegionData(),
-            fleets: [
-              Fleet(
-                id: 'f1',
-                ownerId: 'p1',
-                seaZoneId: 'sea1',
-                regionId: ow,
-                mission: FleetMission.blockade,
-                targetProvinceId: '$ow|P2',
-                shipTypeIds: const ['carrack'],
-              ),
-            ],
-          ),
-          players: const [
-            Player(id: 'p1', displayName: 'A', isHuman: true),
-            Player(id: 'p2', displayName: 'B', isHuman: true),
           ],
           diplomacyRelations: [
             DiplomacyRelation(
@@ -231,45 +146,24 @@ void registerDiplomacyVictoryCoreTests() {
       test(
         'naval interception phase runs when two at-war fleets in same zone',
         () {
-          final topology = MapTopology(
-            nodes: const [
-              TopologyNode(
-                id: 'sea1',
-                regionId: 'oldWorld',
-                type: TopologyNodeType.seaZone,
+          final topology = turnTestOwSeaZoneTopology();
+          final game = turnTestOwGame(
+            provinces: const [],
+            fleets: [
+              Fleet(
+                id: 'fleet_p1',
+                ownerId: 'p1',
+                seaZoneId: 'sea1',
+                regionId: kRegionOldWorld,
+                shipTypeIds: ['carrack'],
               ),
-            ],
-            edges: const [],
-          );
-          final game = Game(
-            id: 'g1',
-            worldState: WorldState(
-              turnState: const TurnState(
-                phase: TurnPhase.orders,
-                turnNumber: 0,
+              Fleet(
+                id: 'fleet_p2',
+                ownerId: 'p2',
+                seaZoneId: 'sea1',
+                regionId: kRegionOldWorld,
+                shipTypeIds: ['fluyte'],
               ),
-              oldWorld: const RegionData(),
-              newWorld: const RegionData(),
-              fleets: [
-                Fleet(
-                  id: 'fleet_p1',
-                  ownerId: 'p1',
-                  seaZoneId: 'sea1',
-                  regionId: 'oldWorld',
-                  shipTypeIds: ['carrack'],
-                ),
-                Fleet(
-                  id: 'fleet_p2',
-                  ownerId: 'p2',
-                  seaZoneId: 'sea1',
-                  regionId: 'oldWorld',
-                  shipTypeIds: ['fluyte'],
-                ),
-              ],
-            ),
-            players: const [
-              Player(id: 'p1', displayName: 'A', isHuman: true),
-              Player(id: 'p2', displayName: 'B', isHuman: true),
             ],
             diplomacyRelations: [
               DiplomacyRelation(
@@ -292,19 +186,9 @@ void registerDiplomacyVictoryCoreTests() {
       );
 
       test('full turn with buildWork applies work order', () {
-        final topology = MapTopology(
-          nodes: [
-            const TopologyNode(
-              id: 'P1',
-              regionId: 'oldWorld',
-              type: TopologyNodeType.province,
-            ),
-          ],
-          edges: [],
-        );
-        const ow = 'oldWorld';
-        const provinceId = 'oldWorld|P1';
-        const tileKey = 'oldWorld|P1|0|0';
+        final topology = turnTestOwSingleProvinceTopology();
+        final provinceId = turnTestOwProvinceId('P1');
+        final tileKey = '$provinceId|0|0';
         final unit = Unit(
           id: 'u1',
           type: kUnitTypeExplorer,
@@ -312,32 +196,13 @@ void registerDiplomacyVictoryCoreTests() {
           locationProvinceId: provinceId,
           tileKey: tileKey,
         );
-        final game = Game(
-          id: 'g1',
-          worldState: WorldState(
-            turnState: const TurnState(phase: TurnPhase.orders, turnNumber: 0),
-            oldWorld: RegionData(
-              provinces: [
-                Province(id: provinceId, regionId: ow, ownerId: 'p1'),
-              ],
-              units: [unit],
-            ),
-            newWorld: const RegionData(),
-          ),
+        final game = turnTestOwGame(
+          provinces: [
+            Province(id: provinceId, regionId: kRegionOldWorld, ownerId: 'p1'),
+          ],
+          units: [unit],
           players: const [Player(id: 'p1', displayName: 'P1', isHuman: true)],
         );
-        final tileMapByRegion = {
-          ow: TileMapResult(
-            width: 1,
-            height: 1,
-            grid: const [
-              ['P1'],
-            ],
-            terrainGrid: [
-              [TerrainType.hills],
-            ],
-          ),
-        };
         final orders = Orders(
           workOrdersByPlayerId: {
             'p1': [
@@ -353,7 +218,7 @@ void registerDiplomacyVictoryCoreTests() {
           game: game,
           topology: topology,
           orders: orders,
-          tileMapByRegion: tileMapByRegion,
+          tileMapByRegion: turnTestSingleTileOwMap('P1'),
           extractedByPlayerId: const {},
           defaultAssignments: const [],
         );
