+++ tool/ct_repo_lint_manifest.yaml
@@ -744,10 +744,17 @@ rules:
 
   - rule_id: repo.turn_test_support_loc
     group: testing
-    title: "colonizethis_turn test/support physical LOC must stay at or below ratchet ceiling (Refs #4039)"
+    title: "colonizethis_turn test/support physical LOC must stay at or below ratchet ceiling (Refs #4039, #4168)"
     spec: SPEC/program/repo-lint.md
     runner: dart
     script: tool/check_turn_test_support_loc.dart
+
+  - rule_id: repo.turn_test_support_region_literals
+    group: testing
+    title: colonizethis_turn test/support must not hard-code bare oldWorld/newWorld literals (Refs #4168)
+    spec: SPEC/program/repo-lint.md
+    runner: dart
+    script: tool/check_turn_test_support_region_literals.dart
 
   - rule_id: repo.turn_lib_file_size
     group: structure
